clean, code cleaning and ray is_forward
